+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Moved `JSON::Validator.absolutize_ref` and the ref manipulating code in
   `JSON::Schema::RefAttribute` into `JSON::Util::URI`
 - Deprecated `JSON::Validator#validator_for` in favor of `JSON::Validator#validator_for_uri`
+- Deprecated `JSON::Validator.validate2` in favor of `JSON::Validator.validate!`
 
 ## [2.7.0] - 2016-09-29